feat: Table View + fix

+++ database/migrations/2025_07_07_175010_create_time_entries_table.php
@@ -17,7 +17,7 @@ return new class extends Migration
             $table->date('day');
             $table->time('started_at');
             $table->time('ended_at');
-            $table->integer('duration_hours')->nullable();
+            $table->decimal('duration_hours', 3, 2)->nullable();
             $table->text('notes')->nullable();
             $table->timestamps();
         });
